[rom_ext] Add the flash exception hander
Refactor the flash exception handler used in the ROM to be usable in
both the ROM and ROM_EXT.
