+++ sw/device/silicon_creator/lib/BUILD
@@ -125,6 +125,11 @@ cc_library(
     ],
 )
 
+filegroup(
+    name = "flash_exc_handler",
+    srcs = ["flash_exc_handler.S"],
+)
+
 cc_library(
     name = "epmp_defs",
     hdrs = [
