+++ sw/device/silicon_creator/rom_ext/BUILD
@@ -77,6 +77,27 @@ cc_test(
     ],
 )
 
+cc_library(
+    name = "rom_ext_isrs",
+    srcs = [
+        "rom_ext_isrs.c",
+        "//sw/device/silicon_creator/lib:flash_exc_handler",
+    ],
+    hdrs = ["rom_ext_isrs.h"],
+    defines = [
+        "INTERRUPT_HANDLER=rom_ext_interrupt_handler",
+    ],
+    target_compatible_with = [OPENTITAN_CPU],
+    deps = [
+        "//hw/top_earlgrey/ip_autogen/flash_ctrl:flash_ctrl_c_regs",
+        "//hw/top_earlgrey/sw/autogen:top_earlgrey",
+        "//sw/device/lib/base:csr",
+        "//sw/device/lib/base:macros",
+        "//sw/device/silicon_creator/lib:dbg_print",
+        "//sw/device/silicon_creator/lib:error",
+    ],
+)
+
 ld_library(
     name = "ld_common",
     includes = ["rom_ext_common.ld"],
@@ -126,6 +147,7 @@ ld_library(
         deps = [
             ":rom_ext_boot_policy",
             ":rom_ext_boot_policy_ptrs",
+            ":rom_ext_isrs",
             ":rom_ext_manifest",
             "//hw/ip/otp_ctrl/data:otp_ctrl_c_regs",
             "//hw/ip/sram_ctrl/data:sram_ctrl_c_regs",
