+++ sw/device/silicon_creator/rom_ext/rom_ext_common.ld
@@ -111,6 +111,20 @@ SECTIONS {
   } > rom_ext_flash
 
   /**
+   * ROM_EXT ISRs text section, containing the exception handler entry/exit logic
+   * used for handling load access faults due to flash ECC errors during boot.
+   *
+   * This is placed next to the ".shutdown" section.
+   */
+  .isrs : ALIGN(4) {
+    *(.isrs)
+    *(.isrs.*)
+
+    /* Ensure section end is word-aligned. */
+    . = ALIGN(4);
+  } > rom_ext_flash
+
+  /**
    * Shutdown text section, containing shutdown function(s).
    *
    * This must be the last executable section in the ROM_EXT flash image.
